Change expand ratio to int to be compatible with flexbox

+++ server/src/com/vaadin/shared/ui/orderedlayout/AbstractOrderedLayoutState.java
@@ -33,6 +33,6 @@ public class AbstractOrderedLayoutState extends AbstractLayoutState {
 
         public int alignmentBitmask = AlignmentInfo.TOP_LEFT.getBitMask();
 
-        public float expandRatio = 0.0f;
+        public int expandRatio = 0;
     }
 }
